Fix to a bug in the gulag code (#20)
In the case that the bot failed to send a dm to a user it would just
quit. This would result in a user stuck in the gulag. Some causes for
this included the user in question blocking the bot or the user being a
bot as well and not being able to receive dms.